+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.2.0](https://github.com/karma-yrb/PIMMS-ateliers-Word/compare/v1.1.0...v1.2.0) (2026-05-04)
+
+
+### Features
+
+* **release:** automate header version bump and releases page ([e1df579](https://github.com/karma-yrb/PIMMS-ateliers-Word/commit/e1df5795829862061ca086e615a982b46ff4e43c))
+
+
+### Bug Fixes
+
+* **release:** avoid prebump stdout version override ([27bf455](https://github.com/karma-yrb/PIMMS-ateliers-Word/commit/27bf455a80d92dbcf381b4b22492811093e3f398))
+
 ## [1.1.0](///compare/v1.0.0...v1.1.0) (2026-04-30)